requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cssselect2==0.7.0
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
Django==4.2.7
django-celery==3.1.17
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-crispy-forms==2.1
django-email-verification==0.3.3
django-environ==0.11.2
django-htmx==1.17.2
django-mathfilters==1.0.0
django-timezone-field==6.1.0
Faker==25.2.0
flower==2.0.1
fonttools==4.51.0
html5lib==1.1
humanize==4.9.0
idna==3.7
kombu==5.3.7
netaddr==1.2.1
packaging==24.0
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycparser==2.22
pydyf==0.10.0
PyJWT==2.8.0
pyphen==0.15.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
requests==2.31.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.5.0
stripe==9.6.0
tinycss2==1.3.0
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
yookassa==3.1.0
zopfli==0.2.3
